add dependency checks and some refactoring, version 0.5.3

+++ src/main/java/pl/psnc/dl/ege/utils/DataBuffer.java
@@ -325,11 +325,9 @@ public class DataBuffer
 		 */
 		public void deleteFile()
 		{
-			if (!os.isInMemory()) {
-				if (os.getFile().exists()) {
+			if (!os.isInMemory() && os.getFile().exists()) {
 					LOGGER.debug("Removing tmp file : " + os.getFile());
 					os.getFile().delete();
-				}
 			}
 		}
